fix round problem
The rounding is the mathematical definition of rounding which means for
each digit to be rounded, if it is 5 or more, it will be rounded up;
otherwise, it is truncated. For example, rounding 0.215 to two digits
should result in 0.22. However, using the round(0.25 , 2) function in
python 2.7 yields 0.21.

So, you can either implement your own rounding function, or import
decimal package in python.
Keep in mind that if you use decimal() function straight up, it will
give you the same result as round() function, which is NOT what we want.
Below is the correct way of using the decimal package for correct
rounding results:

>>>from decimal import Decimal
>>> a = 0.215
>>> Decimal(str(a)).quantize(Decimal('.01'))
Decimal('0.22')